+++ src/test/java/edu/illinois/library/cantaloupe/resource/iiif/v2/ImageResourceTest.java
@@ -408,15 +408,15 @@ public class ImageResourceTest extends ResourceTest {
     }
 
     @Test
-    void testGETMoreThanMaxPixels() {
+    void testGETMoreThanMaxPixelsWithFullSizeArgument() {
         URI uri = getHTTPURI("/" + IMAGE + "/full/full/0/color.png");
-        tester.testMoreThanMaxPixels(uri);
+        tester.testForbiddingMoreThanMaxPixels(uri);
     }
 
     @Test
-    void testGETMaxPixelsIgnoredWhenStreamingSource() {
-        URI uri = getHTTPURI("/" + IMAGE + "/full/full/0/color.jpg");
-        tester.testMaxPixelsIgnoredWhenStreamingSource(uri);
+    void testGETMoreThanMaxPixelsWithMaxSizeArgument() throws Exception {
+        URI uri = getHTTPURI("/" + IMAGE + "/full/max/0/color.png");
+        tester.testDownscalingToMaxPixels(uri, 64, 56, 1000);
     }
 
     @Test